Created strong names for all assemblies.  The password for the key is "mistral". : Removed : the reference to Blowfish_net Blowfish_net.dll out of lib Habanero.Base.BlowfishCrypter Habanero.Bo.PasswordBlowfish Habanero.Test.Bo.TestPasswordBlowFish
This was because of Blowfish_NET's license - it's not available for commercial purposes.

Removed: 
reference to MyTimePicker
MyTimePicker.dll out of lib
Habanero.Ui.Forms.TimePickerMapper
Habanero.Test.Ui.BoControls.TestTimePickerMapper

MyTimePicker is not strong-named and thus the project wouldn't build.  I was able to rebuild both Noogen.WinForms.dll and nmock.dll with strong names.

Also managed to get the InternalsVisibleTo attribute working, so now the tests in Habanero.Test.Bo can access internal stuff in Habanero.Bo, so the tests that are in there can be moved out.  Will do this next.