(Merge from Branch 2.2: r4881-4919): Added ObjectID to IPrimaryKey to uniquely identify a BO. Added validation and corrections to Primary Keys. Changed Signature of PropertyUpdated on IBusinessObject to use BOPropUpdatedEventArgs. Added code file headers Fixed some code file headers Updating documentation. Fixed issues in Habanero Solutions and projects. Changed BusinessObjectManager To always use the ObjectID (i.e. a guid assigned to the object on creation or on loading from the database) and to never use the ToString of the Primary Key. The behaviour will be identical in the case of Business objects that use Guid ID's but will be different for any objects that use a Primary key that uses properties that are not GUID's.
Added rules to ClassDef loading that detects and fixed a configuration error where the BusinessObject.PrimaryKey is set to be an isObjectID but the PropDef is not set to be WriteNew and Compulsory the ClassDefLoader automatically sets this up.
The reason for not throwing an error is to prevent old project from causing problems.
Changed DataSetProvider To always use the ObjectID (i.e. a guid assigned to the object on creation or on loading from the database) and to never use the ToString of the Primary Key. The behaviour will be identical in the case of Business objects that use Guid ID's but will be different for any objects that use a Primary key that uses properties that are not GUID's.