Created Created IBOSelector with the following inherited interfaces IBOComboBoxSelector.cs, IBOListBoxSelector.cs, IBOCollapsiblePanelSelector.cs and implemented this interface on IGridBase
Added A GroupBoxGroupControl (This is a Control that allows you to have a group of Collapsible Panels on a Panel.

Created an IGroupControl this is a Control that manages a Group of other controls. The IGroupControl is used by the Panel Builder to Build a Grouping Panel i.e. a Group of Panels built by the PanelBuilder. The Panel Builder can now create the Panel to use a Tab, CollapsiblePanel or a GroupBox as its grouping instead of only using the Tab.