Fixed Bug in ClassDef.GetPropertyType where the property type was being returned as object when the PropDef was a LookupList. This was causing any orderBy for a LookupList being returned incorrectly. This is only ever relevant where a lookup is being done for an int or a string Property e.g. A SimpleLookupList. And you want to sort by the underlying property value. To facilitate this it was required that changes where made to DataSetProvider and GridInitialiser so as to convert LookupList props to type object.
Added ClassName to PropDef. 

Added Capability to HabaneroMenu where you can set the ManagedControl i.e. the control that is being managed by the menu item. This managed control is then wrapped by a default IControlManager and the remainder of the menu stays the same.

Implemented the Methods used for Paging e.g. ItemsPerPage on DataGridViewWin.cs. This is to prevent a Control that is set up in VWG with paging from giving not implemented exception in win.