Fixed #700 (Invalid creation of criteria from string when joining over another table) Changed the JoinToSource method in Source to create a shallow clone of the Source being joined to.  This is so that subsequent changes to the source added don't interfere with other source trees which may be using this source object.  Added a test in Source to indicate this bug (Test_MergeTwoSources_CheckThatSecondMergeDoesntAffectSourceOfFirstField). The bug was this:  I had a criteria string as follows:
((Request.ReceivedDate >= '01 Apr 2010 00:00:00') and (Request.ReceivedDate <= '01 May 2010 00:00:00')) and (Request.Institution.Name = 'My Hospital')

And at the point where the collection of objects was loaded the string had changed to:

{((Allegation.Request.Institution.ReceivedDate >= '2010/04/01 00:00:00') AND (Allegation.Request.ReceivedDate <= '2010/05/01 00:00:00')) AND (Allegation.Request.Institution.Name = 'My Hospital')}

Note the Institution field added on there, because the merging of the third field (the Institution.Name field) affected the source of the first field. Using a shallow clone alleviates this as the queryfield's source is disconnected from the criteria's source (where before they were the same object).