Fixed #1708 (Xml export methods on BusinessObject and DataStoreInMemory are not used or correct) Completed #1709 (DataStoreInMemoryXmlWriter/Reader is difficult to get to read to/from a string)
Marked the following methods on BusinessObject as obsolete:
- XmlSchema GetSchema()
- void ReadXml(XmlReader reader)
- void WriteXml(XmlWriter writer)

Marked the following methods on DataStoreInMemory as obsolete:
- void SaveToXml(string fullFileName)
- void LoadFromXml(string fullFileName, Type typeToLoad)

These methods will probably be gone in 2.7 and later. Rather use the DataStoreInMemoryXmlWriter to write BusinessObjects to an xml stream and DataStoreInMemoryXmlReader to read them.

Added a WriteToString method on DataStoreInMemoryXmlWriter
Added a ReadFromString method on DataStoreInMemoryXmlReader

Small rename of local variable in RSAPasswordCrypter